PP-10004 common testing harness ARM64 support
- package `jnr-ffi` dependency, this dependency enables access to native libraries outside of the JVM (in this case, the docker client on ARM64)
- added abstract helper class to DRY up the Postgres JUnit testing rules/extensions
- overloaded the PostgresContainer constructor to allow user specified image tag
